add archive to plan versions button (#506)
