+++ frontend/src/components/base/PageLayout.tsx
@@ -12,7 +12,7 @@ export const PageLayout = ({
   children,
   hasBackButton,
   backButton,
-  aboveTitle,
+  aboveTitle = true,
   ...props
 }: PageLayoutProps) => {
   return (
@@ -29,9 +29,7 @@ export const PageLayout = ({
       <div className="mx-10 mt-16">
         <div className="flex items-center justify-between mb-6">
           {props.title ? (
-            <h1
-              className={hasBackButton ? "font-main  mx-10" : "font-main"}
-            >
+            <h1 className={hasBackButton ? "font-main  mx-10" : "font-main"}>
               {hasBackButton && aboveTitle && backButton}
               <div className={hasBackButton ? "mt-12" : ""}>{props.title}</div>
             </h1>
